move the permute_config_gen to a central location
