+++ hls4ml/backends/oneapi/passes/reshaping_templates.py
@@ -175,40 +175,6 @@ transpose_task_sequence_template = 'task_sequence<nnet::transpose_stream<{input_
 transpose_include_list = ['nnet_utils/nnet_transpose.h', 'nnet_utils/nnet_transpose_stream.h']
 
 
-def permute_config_gen(name: str, shape: tuple[int, ...], perm: tuple[int, ...]):
-    """
-    Generate a configuration string for a permute operation. Operates by mapping the output index to input input index by:
-     - unravel the output index
-     - map each dimension to the corresponding stride in the input tensor, sum
-    The operation can be expressed as:
-
-    new_shape = tuple(shape[i] for i in perm)
-    strides = np.cumprod((shapes[1:] + (1,))[::-1])[::-1]
-    perm_strides = [strides[i] for i in perm]
-    out[index] = inp[np.dot(np.unravel_index(index, new_shape), perm_strides)]
-
-    Args:
-        name (str): The name of the configuration.
-        shape (tuple[int, ...]): The shape of the input tensor.
-        perm (tuple[int, ...]): The permutation of the dimensions.
-
-    Returns:
-        str: The formatted configuration string for the permute operation.
-    """
-    new_shape = tuple(shape[i] for i in perm)
-    strides = np.cumprod((shape[1:] + (1,))[::-1])[::-1]
-    perm_strides = tuple(int(strides[i]) for i in perm)
-    return transpose_config_template.format(
-        dims=len(shape),
-        N=np.prod(shape),
-        from_shape=', '.join(str(x) for x in shape),
-        perm=', '.join(str(x) for x in perm),
-        perm_strides=', '.join(str(x) for x in perm_strides),
-        to_shape=', '.join(str(x) for x in new_shape),
-        config_name=name,
-    )
-
-
 class TransposeConfigTemplate(LayerConfigTemplate):
     def __init__(self):
         super().__init__(Transpose)
@@ -218,7 +184,17 @@ class TransposeConfigTemplate(LayerConfigTemplate):
         shape = tuple(node.get_input_variable().shape)
         perm = tuple(node.get_attr('perm'))
         name = f'config{node.index}'
-        return permute_config_gen(name, shape, perm)
+
+        new_shape, perm_strides = node.model.config.backend.permute_config_gen(name, shape, perm)
+        return transpose_config_template.format(
+            dims=len(shape),
+            N=np.prod(shape),
+            from_shape=', '.join(str(x) for x in shape),
+            perm=', '.join(str(x) for x in perm),
+            perm_strides=', '.join(str(x) for x in perm_strides),
+            to_shape=', '.join(str(x) for x in new_shape),
+            config_name=name,
+        )
 
 
 class TransposeFunctionTemplate(FunctionCallTemplate):
